[ADD] Hook to use the width of the window dynamically. The components were also updated.
